Initial Commit
This initial commit implements a simple circle button and layout manager
for a Rock-Paper-Scissors game. This is just a simple prototype for
introductory game networking -- which is not yet implemented. UDP for
non-blocking communication seems like a good approach.

The current implementation is admittedly awkward, the purpose of
creating the repository is not just version control, but also to track
further development if I even want to polish it.